# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lease-Contract Kiosk API
  version: 0.2.0
  description: |
    - 부동산 등기변동 모니터링 · 전자계약 백엔드  
    - **stub 모드** ➀ 등기부 API, ➁ 메일, ➂ S3 를 모두 오프라인으로 대체  
    - `Authorization: Bearer <JWT>` 헤더 필요 (로그인 제외)
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []          # 모든 경로 기본적용
...
paths:
...
  /contracts:
    post:
      tags: [Contracts]
      summary: 계약 초안(DRAFT) 생성
      operationId: createContract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractDraft' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Contract' } } } }
  /contracts/{id}/pdf:
    post:
      tags: [Contracts]
      summary: 계약서 PDF 생성 + (로컬/S3) 업로드
      operationId: generateContractPdf
      parameters:
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: PDF 업로드 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string, description: S3 (or local) 키 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ContractId:
      in: path
      name: id
      required: true
      schema: { type: string }
...
  schemas:
...
    ContractDraft:
      type: object
      required: [tenantName, landlordName, monthlyRent]
      properties:
        tenantName:   { type: string }
        landlordName: { type: string }
        monthlyRent:  { type: integer, format: int32 }
    Contract:
      allOf:
        - $ref: '#/components/schemas/ContractDraft'
        - type: object
          required: [_id, status]
          properties:
            _id:    { type: string }
            status:
              type: string
              enum: [DRAFT, SIGNING, COMPLETE]